Add log-file option to capture the actions of the monitor.
Improve robustness around spawning and monitoring the tracer process by
making the list manipulations signal safe (we block SIGCHLD while
manipulating the list) and by adding a separate tracer-interval
parameter to establish a deadline for extant tracers to complete their
work.

Resolve an issue where the monitor may consider a tracer as not being in
the running state and terminate the to-be-traced child out from under
the tracer.

refs #8886 (spent 1)