fix unwinding through syscalls such as kill(2)
We were assuming that the CIE for an FDE was always the one that
preceeded the FDE, but this is not the case.

A nice side effect of this change is that it accelerates resolution of
FDEs when multiple functions are present in the same module, but it does
increase the memory cost.

refs #1218 (spent 3)