resolve an issue on 32-bit systems.
We treat pointers as 64-bit througout the code but had some places
that assumed 32-bit.  This lead to a crash in printf on a 32-bit linux
system.

Resolved this by making sure we're not mixing void* with gimli_addr_t
and adding appropriate accessors.

Also discovered that using pread/pwrite is dangerous on a 32-bit system
in general, as the top of stack ends up being represented by a negative
off_t offset, and leads to failure to correctly interrogate the target
process.

Moving to pread64 and pwrite64 resolves this issue.

This patch is mostly cast removal or fixing.
I have compiled and run this test:

{{{
LD_LIBRARY_PATH=.libs ./monitor --debug=1 --detach=0 \
   --glider=.libs/glider --run-once=1 ./wedgie
}}}

on a 64-bit Linux system, building with `CC="gcc -m32"` to verify the
32-bit compilation, and a 64-Solaris system doing the same.

refs #11187 (spent 4)