try harder to detect a stopped process on Linux.
It feels hacky for sure; we give the target 5 seconds to stop, and also
inspect `/proc/pid/status` to see if the target is in the stopped state
(this handles some edges where the target stops before it becomes our
child).  If it has still not stopped, send a SIGSTOP and give it another
5 seconds.

Putting this on the same branch as refs #11187 to make it easier to test
out the Gimli changes overall.

refs #11137 (spent 0.5)