Add a configure option to link against libunwind
This is wired up for linux only at the moment; Darwin has a mach specific
implementation that we could probably adapt to in the future, and I just haven't
gotten around to trying this with Illumos.  Also: Solaris systems tend to be
very good about keeping frame pointers and don't typically need DWARF unwinding.

If you request libunwind and it is not present, configure will not fail; you'll
just end up using the built-in unwinder.

Why?  I've seen some issues with the built-in unwinder and haven't had
time to delve in and resolve them.  They manifest as apparent "invalid read"
attempts when dumping out the full stack trace.  What is actually happening
is that the unwind has lost track of the registers and cannot properly
locate the variables in the frames.