Allow for multiple tracer modules to be declared in an image
Useful for situations where you have statically linked multiple
libraries that have gimli tracer implementations.

This uses a named section in the object file that holds a set of
string pointers.

The linker will stick this together, and it is ok and expected that
there can be duplicate entries for a given module (for example:
two different source files can declare that they are traced by the
same module).